prototypes/receiver/requirements.txt
async-timeout==2.0.1
python-vlc==3.0.102
websockets==4.0.1